Inspiration window

+++ backend/utils/adjancency_db_list.js
@@ -3,7 +3,7 @@ import { Semaphore } from 'async-mutex';
 
 // tableName = just the tableName
 // fields = ,fieldName,otherFieldName (every fieldname should be started with a ,)
-// isLeaf => receives the full node info and should decide if it is a leaf (async function)
+// isLeaf => receives the full node info and should decide if it is a leaf (async function) (only for optimization, can also always return false)
 export function CreateTableInfo(tableName, fields, isLeaf) {
     return { tableName: tableName, tableFields: fields, isLeaf };
 }
@@ -15,23 +15,24 @@ export function GetSemaphore(tableName) {
     }
     return semaphores[tableName];
 }
-async function ResolveConflicts(conflicts, resolveConflicts, movingID, { tableInfo, modificationAction, onNodeDelete }) {
+async function ResolveConflicts(conflicts, resolveConflicts, movingID, { tableInfo, modificationAction, onNodeDelete, onStartModifying, onIDChange }) {
     let toBeKept = [];
     let parentID = undefined;
     if (conflicts.length > 0) {
         // Go through all the conflicts and look at their respective matches in the override array
         conflicts.forEach((conflict) => {
-            if (resolveConflicts[conflict.child.id] == undefined) { return 'Incorrect override array'; }
-            if (resolveConflicts[conflict.child.id] === 'replace') {
-                onNodeDelete(conflict.with);
+            if (resolveConflicts !== true && resolveConflicts[conflict.child.id] == undefined) { return 'Incorrect override array'; }
+            if (resolveConflicts === true || resolveConflicts[conflict.child.id] === 'replace') {
+                onNodeDelete(conflict.with, conflict.child);
             } else if (resolveConflicts[conflict.child.id] === 'ignore') { parentID=true }
             else { return 'Incorrect override array'; }
         });
         if (parentID) { parentID = (await DB.GetNode(movingID, tableInfo)).parent_id; }
+        if(onStartModifying != undefined) await onStartModifying();
 
         let toBeRemoved = [];
         conflicts.forEach((conflict) => {
-            if (resolveConflicts[conflict.child.id] === 'replace') {
+            if (resolveConflicts === true || resolveConflicts[conflict.child.id] === 'replace') {
                 // Remove the file at the new location
                 toBeRemoved.push(DB.DeleteNode(conflict.with.id, tableInfo));
             } else if (resolveConflicts[conflict.child.id] === 'ignore') {
@@ -48,47 +49,56 @@ async function ResolveConflicts(conflicts, resolveConflicts, movingID, { tableIn
     // Replace the nodes that needed to be kept:
     if (toBeKept.length > 0) {
         for (let i = 0; i < toBeKept.length; i++) {
-            if (!(await DB.CreateNodeAtPath(parentID, toBeKept[i].path, DB.GetExtraFieldsOfNode(toBeKept[i], tableInfo), tableInfo))) {
+            const newID = await DB.CreateNodeAtPath(parentID, toBeKept[i].path, DB.GetExtraFieldsOfNode(toBeKept[i], tableInfo), tableInfo);
+            if (!newID) {
                 console.error('Failed to insert node ids');
-                onNodeDelete(toBekept[i]);
+                onNodeDelete(toBekept[i], undefined);
             }
+            onIDChange(toBeKept[i].id, newID, toBeKept[i]);
         }
     }
 
     return false;
 }
 
-export async function CreateNode(parentId, otherValues, tableInfo, { onComplete, onError }) {
+export async function CreateNode(parentId, otherValues, tableInfo, { onComplete, onSephamoreActivate, onError }) {
     await GetSemaphore(tableInfo.tableName).runExclusive(async (val) => {
+        if(onSephamoreActivate != undefined) { const result = await onSephamoreActivate(); if(!result) return }
+
         let newId = await DB.CreateNodeReturnID(parentId, otherValues, tableInfo);
-        await onComplete(newId);
+        if(onComplete != undefined) await onComplete(newId);
     }).catch((err) => {
         console.error('Error while adding node: ' + err.stack);
         return onError('Server error');
     });
 }
 
-export async function DeleteNode(id, tableInfo, { onNodeDelete, onComplete, onError }) {
+export async function DeleteNode(id, tableInfo, { onNodeDelete, onSephamoreActivate, onComplete, onError }) {
     await GetSemaphore(tableInfo.tableName).runExclusive(async (val) => {
+        if(onSephamoreActivate != undefined) { const result = await onSephamoreActivate(); if(!result) return }
+
         const node = await DB.GetNode(id, tableInfo);
         if (!(await tableInfo.isLeaf(node))) {
             const children = await DB.GetChildrenOfNode(id, tableInfo);
-            children.forEach((child) => onNodeDelete(child));
+            children.forEach((child) => onNodeDelete(child, undefined));
+            onNodeDelete(node, undefined);
         } else {
-            onNodeDelete(node);
+            onNodeDelete(node, undefined);
         }
 
         await DB.DeleteNode(id, tableInfo);
 
-        await onComplete();
+        if(onComplete != undefined) await onComplete();
     }).catch((err) => {
         console.error('Error while deleting nodes: ' + err.stack);
         return onError('Server error');
     });
 }
 
-export async function RenameNode(id, newName, override, tableInfo, { onNodeDelete, onComplete, onInvalidInput, onError }) {
+export async function RenameNode(id, newName, override, tableInfo, { onNodeDelete, onIDChange, onSephamoreActivate, onComplete, onStartModifying, onInvalidInput, onError }) {
     await GetSemaphore(tableInfo.tableName).runExclusive(async (val) => {
+        if(onSephamoreActivate != undefined) { const result = await onSephamoreActivate(); if(!result) return }
+
         if (override == undefined) return onInvalidInput('Please specify the overrides');
         if (id == null) return onInvalidInput("Can't rename the root");
 
@@ -103,48 +113,58 @@ export async function RenameNode(id, newName, override, tableInfo, { onNodeDelet
         const result = await ResolveConflicts(conflicts, override, id, {
             tableInfo: tableInfo,
             modificationAction: async () => {
-                let failedNodes = await DB.RenameNode(id, newName, tableInfo);
+                let [failedNodes, changedIDs] = await DB.RenameNode(id, newName, tableInfo);
                 failedNodes.forEach((node) => {
-                    onNodeDelete(node);
-                })
+                    onNodeDelete(node, undefined);
+                });
+                changedIDs.forEach(({ from, to, nodeInfo }) => {
+                    onIDChange(from, to, nodeInfo);
+                });
             },
-            onNodeDelete: onNodeDelete
+            onNodeDelete: onNodeDelete,
+            onStartModifying: onStartModifying
         });
         if(result) return onInvalidInput(result);
 
-        await onComplete();
+        if(onComplete != undefined) await onComplete();
     }).catch((err) => {
         console.error('Error while renaming nodes: ' + err.stack);
         return onError('Server error');
     });
 }
 
-export async function MoveNode(id, newParentId, override, tableInfo, { onNodeDelete, onComplete, onInvalidInput, onError }) {
+export async function MoveNode(id, newParentId, override, tableInfo, { onNodeDelete, onIDChange, onSephamoreActivate, onComplete, onStartModifying, onInvalidInput, onError }) {
     await GetSemaphore(tableInfo.tableName).runExclusive(async (val) => {
+        if(onSephamoreActivate != undefined) { const result = await onSephamoreActivate(); if(!result) return }
+
         if (override == undefined) return onInvalidInput('Please specify the overrides');
         if (id == null) return onInvalidInput("Can't move the root");
 
         const currentParentId = (await DB.GetNode(id, tableInfo)).parent_id;
         if (currentParentId === undefined) { console.error('Parent id is undefined'); return onInvalidInput('Incorrect file id'); }
-        if (newParentId == currentParentId) { await onComplete(); return; }
+        if (newParentId == currentParentId) { await onComplete(undefined, false); return; }
 
         const conflicts = await DB.GetNodeMovingConflicts(id, newParentId, tableInfo);
         if (!override && conflicts.length > 0) {
-            return await onComplete({ conflicts: conflicts });
+            return await onComplete({ conflicts: conflicts }, false);
         }
         const result = await ResolveConflicts(conflicts, override, id, {
             tableInfo: tableInfo,
             modificationAction: async () => {
-                let failedNodes = await DB.MoveNode(id, newParentId, tableInfo);
+                let [failedNodes, changedIDs] = await DB.MoveNode(id, newParentId, tableInfo);
                 failedNodes.forEach((node) => {
-                    onNodeDelete(node);
+                    onNodeDelete(node, undefined);
                 })
+                changedIDs.forEach(({ from, to, nodeInfo }) => {
+                    onIDChange(from, to, nodeInfo);
+                });
             },
-            onNodeDelete: onNodeDelete
+            onNodeDelete: onNodeDelete,
+            onStartModifying: onStartModifying
         });
         if(result) return onInvalidInput(result);
 
-        await onComplete();
+        if(onComplete != undefined) await onComplete(undefined, true);
     }).catch((err) => {
         console.error('Error while moving nodes: ' + err.stack);
         return onError('Server error');
